+++ CHANGELOG.md
@@ -20,7 +20,12 @@ Feb 22 2018
 ```
 - add part of the framework for focusUpdateInfo
 - Form css changes (bolded labels, margin)
-- Allow customization of "pay with" button with:
+- Allow customization of "pay with" button in paymentMethodInfo:
+```
+{
+  "payButtonText": "Pay Now"
+}
+```
 
 
 ## 1.1.4